Fixed to compiler warnings on.

+++ source/service/visualization/cxViewGroupData.cpp
@@ -262,7 +262,7 @@ void ViewGroupData::addDataSorted(DataPtr data)
 	DataViewProperties properties = DataViewProperties::createDefault();
 	DataAndViewProperties item(data, properties);
 
-	for (int i=mData.size()-1; i>=0; --i)
+    for (size_t i=mData.size()-1; i>=0; --i)
 	{
 		if (!dataTypeSort(data, mData[i].first))
 		{
